% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixationTime.R
\name{fixation_time}
\alias{fixation_time}
\title{Fixation Time in Wright-Fisher Model with selection}
\usage{
fixation_time(N, alpha, p)
}
\arguments{
\item{N}{Number of alleles in the considered population (population size)}

\item{alpha}{Selection coefficient}

\item{p}{Vector or initial probabilities (values between 0 and 1)}
}
\value{
the time to fixation
}
\description{
This function computes the time to fixation for a given initial probability and fixed selection coefficient alpha
}
