% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixationTime.R
\name{fixation_time0}
\alias{fixation_time0}
\title{Fixation Time to 0 in Wright-Fisher Model with selection}
\usage{
fixation_time0(N, alpha, p)
}
\arguments{
\item{N}{Number of alleles in the considered population (population size)}

\item{alpha}{Selection coefficient}

\item{p}{Vector or initial probabilities (values between 0 and 1)}
}
\value{
the time to fixation to 0
}
\description{
This function computes the expected time to fixation conditional to a fixation to 0 for a given initial probability and fixed selection coefficient alpha
}
