% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixationTime.R
\name{fixation_time_simu}
\alias{fixation_time_simu}
\title{Fixation Time Simulation in Wright-Fisher Model with selection}
\usage{
fixation_time_simu(N, alpha, nbA = floor(N/2))
}
\arguments{
\item{N}{Total number of alleles in the considered population (population size)}

\item{alpha}{Selection coefficient}

\item{nbA}{number of allele A}
}
\value{
a list containing the time to fixation and its type (fixation to 0 or 1)
}
\description{
Simulation of the Wright-Fisher problem with selection coefficient and given initial population in alleles a and A
}
