% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFixationTimeSelection}
\alias{plotFixationTimeSelection}
\title{Fixation time theoretic curve for different selection coefficients}
\usage{
plotFixationTimeSelection(
  N = 100,
  alphas = c(0, 3, 7, 15, 30),
  p = seq(0.01, 1 - 0.01, by = 0.01)
)
}
\arguments{
\item{N}{Number of alleles in the considered population (population size)}

\item{alphas}{Vector of selection coefficients}

\item{p}{Vector or initial probabilities (values between 0 and 1)}
}
\value{
the time to fixation
}
\description{
This function plots the fixation time curves for a given probability vector and different selection coefficients
}
