% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixationTime.R
\name{upper_bound_value}
\alias{upper_bound_value}
\title{Value of the upper bound majoration}
\usage{
upper_bound_value(alpha, k)
}
\arguments{
\item{alpha}{Selection coefficient}

\item{k}{number of elements in the Taylor expansion}
}
\value{
upper bound value
}
\description{
Upper bound majoring the L1 norm between exact and approximate result for fixation time
}
